refactor(ci): extract the backend and frontend job groups
The last two blocks of getCommonJobs move to workflows/groups/. The method drops from 456 to
50 lines and now reads as what it is: Danger JS, the Helm chart, the backend group, the
frontend group, then the validation gate.

Both groups return { jobs, requires } rather than a job list. The gate at the end of the
workflow waits on names accumulated as the blocks run, and that list is emitted as-is, so a
group has to contribute to both. Callers accumulate the two lists in the order they contribute
the groups, which keeps the emitted job list and the requires list stable.

The gate itself stays in the workflow: it needs the complete accumulated list, including the
fix-up that forces validation when only the distribution pom changed.

workflow-pull-requests.ts drops from 961 to 124 lines across the four extractions. Generated
configuration is byte-for-byte identical: all 50 fixtures are untouched.
